refactor: separate gaussian and sort pipelines
