+++ src/render/bindings.wgsl
@@ -28,27 +28,8 @@ struct DrawIndirect {
     base_vertex: u32,
     base_instance: u32,
 }
-struct SortingGlobal {
-    digit_histogram: array<array<atomic<u32>, #{RADIX_BASE}>, #{RADIX_DIGIT_PLACES}>,
-    assignment_counter: atomic<u32>,
-}
+
 struct Entry {
     key: u32,
     value: u32,
 }
-@group(3) @binding(0) var<uniform> sorting_pass_index: u32;
-@group(3) @binding(1) var<storage, read_write> sorting: SortingGlobal;
-@group(3) @binding(2) var<storage, read_write> status_counters: array<array<atomic<u32>, #{RADIX_BASE}>>;
-@group(3) @binding(3) var<storage, read_write> draw_indirect: DrawIndirect;
-@group(3) @binding(4) var<storage, read_write> input_entries: array<Entry>;
-@group(3) @binding(5) var<storage, read_write> output_entries: array<Entry>;
-@group(3) @binding(6) var<storage, read> sorted_entries: array<Entry>;
-
-
-struct GaussianOutput {
-    @builtin(position) position: vec4<f32>,
-    @location(0) @interpolate(flat) color: vec4<f32>,
-    @location(1) @interpolate(flat) conic: vec3<f32>,
-    @location(2) @interpolate(linear) uv: vec2<f32>,
-    @location(3) @interpolate(linear) major_minor: vec2<f32>,
-};
