+++ src/render/sort/radix.wgsl
@@ -10,11 +10,25 @@
     input_entries,
     output_entries,
     sorted_entries,
+    Entry,
 }
 #import bevy_gaussian_splatting::transform::{
     world_to_clip,
     in_frustum,
 }
+
+
+struct SortingGlobal {
+    digit_histogram: array<array<atomic<u32>, #{RADIX_BASE}>, #{RADIX_DIGIT_PLACES}>,
+    assignment_counter: atomic<u32>,
+}
+
+@group(3) @binding(0) var<uniform> sorting_pass_index: u32;
+@group(3) @binding(1) var<storage, read_write> sorting: SortingGlobal;
+@group(3) @binding(2) var<storage, read_write> status_counters: array<array<atomic<u32>, #{RADIX_BASE}>>;
+@group(3) @binding(3) var<storage, read_write> draw_indirect: DrawIndirect;
+@group(3) @binding(4) var<storage, read_write> input_entries: array<Entry>;
+@group(3) @binding(5) var<storage, read_write> output_entries: array<Entry>;
 
 
 struct SortingSharedA {
